Always pass the required parent models when defining breadcrumbs
Always passing in the needed models prevents an error from triggering when moving from a child route to a parent route.

If they aren't passed, Ember implicitly sets them based on the routing tree. However, this can trigger errors in the LinkTo component when the implicit model switches to an `undefined` value and a model is expected.