Cache the result from previous runs and return that if needed
Sometimes the params-for helper is re-invoked when transitioning out of the page where it is defined. This has the potential issue that it tries to access route params that are no longer there, triggering the assertions.

The current change caches previous succesful runs and only shows the assertions if the initial run doesn't yield the needed result.