Add plausible analytics script as built for gelinkt-notuleren
This script allows the specification of two environment variables
through the docker-compose.override.yml.  When these are set correctly,
plausible will be injected and logging will commence.

EMBER_ANALYTICS_APP_DOMAIN}

The domain for which the app should be monitored.

eg: example.abb.vlaanderen.be

EMBER_ANALYTICS_PLAUSIBLE_SCRIPT

Full path to the plausible script on the server.