Implement the deep link functionality
This allows users to be redirected to the correct page based on a resource URI. The resource URI is used to retrieve the RDF type(s) and uuid of a resource which will then be mapped to the correct RedirectHandler class.

The current implementation supports direct linking to "people" and "worship administrative units".