feat: add legal name attribute for all organisations
Add a legal name to organisations next to the preferred name. Users can edit the
legal name of an organisation, except for municipalities and provinces. The name
displayed for organisations throughout the frontend is the legal name if
present, otherwise the value for the preferred name is used. Search for a name
now searches in both the legal name and the preferred name.