fix: relax requirement of founder for newly onboarded unit types
While a founding organisation should be mandatory for OCMW associations and
PEVAs the available business data is incomplete. Strictly enforcing
this rule would mean that units lacking a founder at the time of onboarding
cannot be edited later as their validation will fail. This commit relaxing this
rule such that a founding organisation is only required when creating a new OCMW
association or PEVA.