fix: overwrite removed SharePoint identifier with empty string
Previously the commented lines caused any existing SharePoint identifier to
remain unchanged in the triplestore when the user removed to value in the edit
core data form. This makes it impossible for users to remove previously set
SharePoint identifiers. By not transforming empty string to null the existing
SharePoint identifier will be overwritten by an empty string in the
triplestore. A better fix avoiding the insertion of empty strings is advised for
the future.