refactor: moved classification relationship to organization model
The classification code relation was moved from the administrative unit model to
the organization model and was made polymorphic. A new model was introduced for
representative body classification codes. This allows to treat classification
codes uniformly over all organizations.

Along with this this change all classification getters have been moved to the
organization model as well.