refactor: use "Representative body" instead of "Representative organ"
Previously both terms, body and organ, were used interchangeably. This commit
changes all uses of organ to body so that we have consistent terminology
throughout the frontend.

Body was chosen over organ as it is also used in the backend and we already use
the term "governing **body**" for "bestuurs**orgaan**".