refactor: removed `CLASSIFICATION_CODE` in favour of `CLASSIFICATION`
This commit removes the `CLASSIFICATION_CODE` object and all its uses. All
references to properties in `CLASSIFICATION_CODE` have been replace by
equivalent references to the `id` of the same property in `CLASSIFICATION`.

The `administrative-unit-classification-code-test` was removed as well. These
tests were introduced when initially refactoring the `CLASSIFICATION_CODE`
object to reuse the id value of `CLASSIFICATION`. With the removal of
`CLASSIFICATION_CODE` these tests are no longer needed.