fix: moved worship classification checks to administrative unit model
The form for new organizations uses an administrative-unit record. This means if
the functions are placed in the more specific models they cannot be correctly
used in deciding which fields to show.

More concretely, the fields for "Soort eredienst", "Strekking", and
"Grensoverschrijdend" were not shown when creating worship services. The
get-functions used to trigger displaying these fields were defined in the
worship service model and therefore not available in the form.