fix: use worship administrative unit model when user has worship role
Previously the creation of organizations used only the plain administrative unit
model. Consequently, the validations specific to worship administrative units
were never checked. With this change the `WorshipAdminsitrativeUnit` model is
used when the user has the worship role such that validations specified at this
level are verified as well.

The same problem still exists for the `WorshipService` and
`CentralWorshipService` models, but these model do not enforce additional
mandatory fields. (The crossborder field in `WorshipServices` defaults to `false`.)