refactor: create most specific organisation model in the fly
Instead of instantiating the available models up-front in the router, the
controller now creates to most specific organization model on the fly based on
the classification code selected by the user.

This allows to perform all validations defined on the level of the model that
will eventually be saved to the backend. Furthermore this refactoring fixes a
bug were properties specific to certain classifications were copied irrelevant
of the final classification.