fix: remove empty decision for change event
When a user removed all entries concerning a decision while editing a change
event this change was not properly persisted. As the change event was empty,
`isEmpty` returns true, its changes were simply ignored and the old decision
remained as is.

This behaviour was changed to take explicit actions when a decision for a change
event becomes empty after edits by a user. Specifically:
- the relation between the change event and decision is set to `null`
- the decision record is destroyed to avoid keeping obsolete data in the
  triplestore