feat: use KBO organization's name as first fallback
With the introduction of the different organisation names it was intended to use
the KBO organization's name as a fallback value if there is no legal name
set. At the time the additional name functionality was added the KBO/wegwijs
service was not yet finished and the name variable was used as only fallback.

This commit resolves the leftover TODO and adds the KBO organization's name as
first fallback value for the `abbName` getter.