feat: filter related organizations by membership role (direction)
Add a membership role filter to the view and edit pages for related
organizations. This filter allows to limit the displayed related organizations
to those related by a specific membership in a specific direction.

For the edit related organizations page the value of the filter can only be
adjusted when there are no unsaved edits. If a membership was added or removed
the value of the filter can not be changed until the changes are saved or discarded.