Fix an issue where the change event select options weren't properly loaded sometimes
`store.findAll` doesn't wait for the background reload to finish if the store already contains records of the specific type that is requested. Since we only run the `loadChangeEventTypesTask` on component creation, the full data set is never used.

By adding the `reload` option we ensure that the full dataset is loaded properly.