fix: use model relation to find other local involvements
When explicitly querying the store for local involvements any newly added, thus
unsaved, local involvements are not retrieved. As a result they were not
considered in the financing percentage validation when calculating the sum of
all percentages.

As a result the validation failed, resulting in error message presenting to the
user they:
1. added a new local involvement
2. adjusted the percentages to correctly sum up to 100

By using the model relation all local involvements are considered in the
percentage validation irrelevant of whether they have been stored before saved.