fix: corrected province relationship tests
The previously configured relationships allowed for a province did not reflect
reality. This updates the tests concerning expected relationships to correct
this.

A province can act as founding organisation for:
- APB
- PEVA province

A province can participate in:
- APB

A province can have a generic relation with:
- APB
- municipality
- PEVA province
- OCMW

A province **cannot** have a
- founding organisation
- participant

chore: fix linting