fix: ensure `memberships` is an EmberArray while it can be edited
The logic to check and correct the `member` and `organization` assignments
before saving `memberships` has the unintended side effect of converting this
from an EmberArray to a regular one.  As a result, any edit after an
unsuccessful save will throw a `TypeError`s as the used `pushObject` and
`removeObject` functions are not defined for regular arrays.

To fix this `memberships` is reconverted to an EmberArray before further
processing within the `save` function.