fix: handle "has a relation with" memberships in both directions
In contrast to the other membership roles, the "has a relation with" role has no
clear direction. The choices provided to the user should be limited on whether
the active organisation would be assigned as `member` or `organization`. The
decision of how to assign these is purely technical and should not impact the
user.

The extension of the tests reflects this in that the expected returned
classification codes should be the same irrelevant of whether an organisation is
assigned as `member` or `member`.

The updated membership configuration reflects how we intend to store these kind
of relationships. For our stored data we do want to keep "direction" of the
memberships consistent.