refactor: use selected membership role to assign active organization
Previously the assignment of organizations to a membership with the role "has a
relation with" could be altered by the code at multiple points:
- the `updateMembershipRole` function enforced municipalities and central
  worship services were always assigned as the membership's `organization`; and
- the `save` function enforced in turn that representative bodies and worship
  services are respectively assigned as `member` and `organization`. For central
  worship service the same is implicitly done by the `updateMembershipRole`
  function.

In this commit the responsibility to assign the current organization to the
correct field, `member` or `organization`, is put solely in the
`updateMembershipRole` function. These assignments will no longer be altered by
the `save` function afterwards. The underlying assumption is that the form does
correctly enforce which classifications the other organization can have such
that a user cannot create memberships that are invalid according to the
configurations in `constants/memberships.js`.