refactor: simplified code to get opposite classifications
Filter possible members and organizations up front and return the needed one
based on the
- membership role; and
- the relation the provided organization has in the membership.

This simplifies the code a bit reducing the amount of similar
filter/flatmaps. The disadvantage is that an unnecessary filter/flatmap is
performed for to construct a `members` or `organizations` that is not actually
used.