fix: do not remove persisted elements from `memberships` array
The bug fix in #642 was applies to broadly. The `save` function relies on the
information in the `memberships` array to persist the necessary changes.
Consequently, removing elements from this array results in any changes,
including their deletion, to be forgotten.

We do explicitly remove new, not yet persisted, memberships from the array and
immediately destroy them. Otherwise, they can cause failed validations, as they
might be incomplete, or cause errors being thrown.