fix: when necessary correct involved organizations assignment on save
For "has a relation with" memberships the `updateMembershipRole` does not always
have enough information to correctly assign the current organization as either
`member` or `organization`. This is due to the fact that the 'correct'
assignment depends on what kind of other organization is selected.
For example, if the current organization is a worship service it should be
assigned as
- `member` when the other organization is a central worship service; and
- as `organization` when the other organization is a representative body.
The previous implementation incorrectly always assigned worship services as
`organization`, irrelevant of the classification of the other involved
organization.

This commit corrects that behaviour by
- defaulting to assign the current organization as member for "has a relation
  with" memberships in `updateMembershipRole`; and
- if necessary, swapping the `member` and `organization` assignments in the
  `save` as this function does have sufficient information to determine the
  'correct' assignment.

chore: fixed typos