fix: filter duplicate worship mandatories in route
Some harvested persons assigned to worship mandatories have duplicate entries
for their name. Specifically, their given or family name is stored twice, once
with a language tag and once without. This causes the frontend to weird stuff
when sorting on give or family name.

This workaround filters any duplicate records in the query results such that
each mandatory is only shown once, irrelevant of the duplicate names.

Once the duplicate entries are removed from the data and source of these
duplicates is fixed, this workaround can be removed again.