Remove the dynamic import from the yasgui modifier
This was originally done to split of the large yasgui library from the main bundle and only load it once the modifier is used. Under Embroider this is no longer needed if route splitting is enabled since the modifier itself won't be loaded until it is needed by a route.

This reduces the complexity in the code and makes it easier to maintain.