feat: display all resulting organisations for merger change events
This modifies the view and edit change event forms to display all resulting
organisations instead of just one. This change is done to deal with the
municipalities of Borsbeek and Antwerp merging, resulting in a new district
Borsbeek *and* the already existing municipality Antwerp.

For existing merger change events with only one resulting organisations, all of
them, nothing visually changes.