fix: determine filters on organisations page based on user's role
Previously the shown filters were based on whether the results in the table
contained worship organisations or not. This resulted in undesired behaviour
when there were no resulting organisations, with the shown filters falling back
to the non-worship ones.

This fix changes that behaviour to determine the shown filters based on the role
of the authenticated user. This results in consistently showing the correct
filters as a user's role remains the same within a single session.