Revert "Fix a pagination issue on the organizations details page"
This reverts commit 023507dbd5b7feeb24333dbc3f2370d1c65a7a84.

This commit restored pagination that was broken by a workaround for duplicate
data. The fix applied here has the disadvantage that the numbers underneath the
table might not be accurate in case the original workaround removed duplicate
entries.

After discussing the options with business, it was decided to remove both the
original workaround (next commit) and this pagination fix that was applied on
top of it.