Revert "fix: filter duplicate worship mandatories in route"
This reverts commit 63f450b050cdfd8d0c62554bf443ee58b707f4f0.

This quick fix was applied to filter duplicate entries due to duplicate names in
the data, typically one triple with a language tag and one without. But this
workaround breaks the pagination of the table. Since a permanent data fix is
underway it was decided to revert this workaround to avoid confusing users.