WIP error route experiment
This uses the ember-error-route addon's setup to redirect users to the error route when something goes wrong. The error route can then update the url based on the transition which fixes the issue that the URL is incorrect after a error in the model hooks.

The error route itself needs more work:
- detect the type of error we're dealing with.
- in case of a 404, we should either display the not found message, or consider it a permission error and display a "not allowed" message instead.
- in case of any other error we should display a generic error message.