feat: allow a location to be contained in multiple other locations
The data model specifies that an administrative unit can only have a single
location as its scope of operation. As a consequence new location resources will
have to be created whenever an administrative unit can operate in multiple
municipalities. Therefore, we also need to allow that a single location is
contained within multiple other locations, otherwise it would not be always
possible to create the necessary location resources.

For example, say `UnitOne` has as scope of operation municipalities `A` and
`B`. This requires create a new location resource that contains `A` and
`B`. Furthermore, say `UnitTwo` has as scope of operation municipalities `A` and
`C`. This requires creating a new location resource containing `A` and `C`. If a
location would be allowed to be only contained in one other location that would
not be possible as `A` is already contained in the former location resource.