fix: find appropriate reference region for organizations
The logic to determine the reference region an organization is located in broke
due changing the `locatedWithin` relations from a `belongsTo` to a
`hasMany`. This commit fixes that by searching for a reference region among all
the `locatedWithin` locations instead. This does assume that the reference
regions remain mutually exclusive such that an organisation is located in at
most one reference region.

Furthermore, the template was slightly simplified by removing the duplicate
region entries. The template now relies on a model getter to determine whether
to show the region and always displays the `region` set in the route (instead of
following model relations for municipalities.)

Finally, for municipalities the region is now displayed as a disabled input
field instead of a plain string. This allows to remove one of the region
sections in the template. This is also more consistent with the other fields in
the form.