fix: split requests with too long URI into multiple requests
In rare cases a scope of operation contains a large amount of locations, for
example provinces whose scope is all its municipalities. This can result is
calls to the backend that are too long leading to 414 errors from the identifier
or dispatcher service.

Since this is currently a rare occurrence, we opted to manually split the
requests when a large number, more than 20, of locations are requested.