feat: visually show that some relations cannot be deleted by users
The backend already disallows users from deleting relations between
municipalities and provinces, as the `shared` graph containing this data is
read-only. But the frontend allowed such edits, only for them to be silently
discarded by the authorization service.

This updates the edit related organisations form to disable the remove button on
rows for relationships the user should not be able to delete.